+++ demo/src/main/java/com/bigotes/app/controller/PetController.java
@@ -44,7 +44,6 @@ public class PetController {
     @PostMapping("/save")
     public String savePet(@ModelAttribute("pet") Pet pet) {
         service.save(pet);
-        System.out.println("Pet: " + pet);
         return "redirect:/pet/all";
     }
 
@@ -60,7 +59,6 @@ public class PetController {
         Pet pet = service.findById(id);
         if (pet != null) {
             model.addAttribute("pet", pet);
-            System.out.println("Pet: " + pet);
         } else {
             throw new NotFoundException();
         }
@@ -69,7 +67,12 @@ public class PetController {
 
     @GetMapping("/del/{id}")
     public String deletePet(Model model, @PathVariable("id") int id) {
-        service.deleteById(id);
+        Pet pet = service.findById(id);
+        if (pet != null) {
+            service.deleteById(id);
+        } else {
+            throw new NotFoundException();
+        }
         return "redirect:/pet/all";
     }    
 }
